Created galleryCard component, migrated code to keep page.tsx clean
